rename sql_db to sql
